requirements.txt
amqp==5.1.1
asgiref==3.5.1
astroid==2.11.5
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
billiard==3.6.4.0
black @ git+https://github.com/psf/black@fa6caa6ca8489103d22d23f8f4ae4d3569bb115e
blessings==1.7
bpython==0.22.1
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.6
cryptography==37.0.2
curtsies==0.3.10
cwcwidth==0.1.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
Django==4.0.4
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flake8==4.0.1
greenlet==1.1.2
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.1
jupyter-core==4.10.0
kombu==5.2.4
lazy-object-proxy==1.7.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
mypy-lang==0.5.0
nest-asyncio==1.5.5
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2022.5.2
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.13.8
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pywin32==304
pyxdg==0.28
PyYAML==6.0
pyzmq==22.3.0
requests==2.28.0
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.2.0
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
waitress==2.1.2
wcwidth==0.2.5
windows-curses==2.3.0
wrapt==1.14.1
